+++ com.unity.render-pipelines.core/Runtime/RenderGraph/RenderGraphResourceRegistry.cs
@@ -5,323 +5,6 @@ using UnityEngine.Rendering;
 
 namespace UnityEngine.Experimental.Rendering.RenderGraphModule
 {
-    #region Resource Descriptors
-    // BEHOLD C# COPY PASTA
-    // Struct can't be inherited and can't have default member values
-    // Hence the copy paste and the ugly IsValid implementation.
-
-    /// <summary>
-    /// Texture resource handle.
-    /// </summary>
-    [DebuggerDisplay("Texture ({handle})")]
-    public struct TextureHandle
-    {
-        bool m_IsValid;
-        internal int handle { get; private set; }
-        internal int transientPassIndex { get; private set; }
-        internal TextureHandle(int handle, int transientPassIndex = -1) { this.handle = handle; m_IsValid = true; this.transientPassIndex = transientPassIndex; }
-        /// <summary>
-        /// Conversion to int.
-        /// </summary>
-        /// <param name="handle">Texture handle to convert.</param>
-        /// <returns>The integer representation of the handle.</returns>
-        public static implicit operator int(TextureHandle handle) { return handle.handle; }
-        /// <summary>
-        /// Return true if the handle is valid.
-        /// </summary>
-        /// <returns>True if the handle is valid.</returns>
-        public bool IsValid() => m_IsValid;
-
-        /// <summary>
-        /// Equals Override.
-        /// </summary>
-        /// <param name="obj">Other handle to test against.</param>
-        /// <returns>True if both handle are equals.</returns>
-        public override bool Equals(System.Object obj)
-        {
-            //Check for null and compare run-time types.
-            if ((obj == null) || !this.GetType().Equals(obj.GetType()))
-            {
-                return false;
-            }
-            else
-            {
-                TextureHandle texture = (TextureHandle)obj;
-                return texture.handle == handle && texture.m_IsValid == m_IsValid;
-            }
-        }
-
-        /// <summary>
-        /// GetHashCode override.
-        /// </summary>
-        /// <returns></returns>
-        public override int GetHashCode()
-        {
-            return (handle << 2) ^ (m_IsValid ? 333 : 444);
-        }
-    }
-
-    /// <summary>
-    /// Renderer List resource handle.
-    /// </summary>
-    [DebuggerDisplay("RendererList ({handle})")]
-    public struct RendererListHandle
-    {
-        bool m_IsValid;
-        internal int handle { get; private set; }
-        internal RendererListHandle(int handle) { this.handle = handle; m_IsValid = true; }
-        /// <summary>
-        /// Conversion to int.
-        /// </summary>
-        /// <param name="handle">Renderer List handle to convert.</param>
-        /// <returns>The integer representation of the handle.</returns>
-        public static implicit operator int(RendererListHandle handle) { return handle.handle; }
-        /// <summary>
-        /// Return true if the handle is valid.
-        /// </summary>
-        /// <returns>True if the handle is valid.</returns>
-        public bool IsValid() => m_IsValid;
-    }
-
-    /// <summary>
-    /// Compute Buffer resource handle.
-    /// </summary>
-    [DebuggerDisplay("ComputeBuffer ({handle})")]
-    public struct ComputeBufferHandle
-    {
-        bool m_IsValid;
-        internal int handle { get; private set; }
-        internal ComputeBufferHandle(int handle) { this.handle = handle; m_IsValid = true; }
-        /// <summary>
-        /// Conversion to int.
-        /// </summary>
-        /// <param name="handle">Compute Buffer handle to convert.</param>
-        /// <returns>The integer representation of the handle.</returns>
-        public static implicit operator int(ComputeBufferHandle handle) { return handle.handle; }
-        /// <summary>
-        /// Return true if the handle is valid.
-        /// </summary>
-        /// <returns>True if the handle is valid.</returns>
-        public bool IsValid() => m_IsValid;
-    }
-
-    /// <summary>
-    /// The mode that determines the size of a Texture.
-    /// </summary>
-    public enum TextureSizeMode
-    {
-        ///<summary>Explicit size.</summary>
-        Explicit,
-        ///<summary>Size automatically scaled by a Vector.</summary>
-        Scale,
-        ///<summary>Size automatically scaled by a Functor.</summary>
-        Functor
-    }
-
-    /// <summary>
-    /// Subset of the texture desc containing information for fast memory allocation (when platform supports it)
-    /// </summary>
-    public struct FastMemoryDesc
-    {
-        ///<summary>Whether the texture will be in fast memory.</summary>
-        public bool inFastMemory;
-        ///<summary>Flag to determine what parts of the render target is spilled if not fully resident in fast memory.</summary>
-        public FastMemoryFlags flags;
-        ///<summary>How much of the render target is to be switched into fast memory (between 0 and 1).</summary>
-        public float residencyFraction;
-    }
-
-    /// <summary>
-    /// Descriptor used to create texture resources
-    /// </summary>
-    public struct TextureDesc
-    {
-        ///<summary>Texture sizing mode.</summary>
-        public TextureSizeMode sizeMode;
-        ///<summary>Texture width.</summary>
-        public int width;
-        ///<summary>Texture height.</summary>
-        public int height;
-        ///<summary>Number of texture slices..</summary>
-        public int slices;
-        ///<summary>Texture scale.</summary>
-        public Vector2 scale;
-        ///<summary>Texture scale function.</summary>
-        public ScaleFunc func;
-        ///<summary>Depth buffer bit depth.</summary>
-        public DepthBits depthBufferBits;
-        ///<summary>Color format.</summary>
-        public GraphicsFormat colorFormat;
-        ///<summary>Filtering mode.</summary>
-        public FilterMode filterMode;
-        ///<summary>Addressing mode.</summary>
-        public TextureWrapMode wrapMode;
-        ///<summary>Texture dimension.</summary>
-        public TextureDimension dimension;
-        ///<summary>Enable random UAV read/write on the texture.</summary>
-        public bool enableRandomWrite;
-        ///<summary>Texture needs mip maps.</summary>
-        public bool useMipMap;
-        ///<summary>Automatically generate mip maps.</summary>
-        public bool autoGenerateMips;
-        ///<summary>Texture is a shadow map.</summary>
-        public bool isShadowMap;
-        ///<summary>Anisotropic filtering level.</summary>
-        public int anisoLevel;
-        ///<summary>Mip map bias.</summary>
-        public float mipMapBias;
-        ///<summary>Textre is multisampled. Only supported for Scale and Functor size mode.</summary>
-        public bool enableMSAA;
-        ///<summary>Number of MSAA samples. Only supported for Explicit size mode.</summary>
-        public MSAASamples msaaSamples;
-        ///<summary>Bind texture multi sampled.</summary>
-        public bool bindTextureMS;
-        ///<summary>Texture uses dynamic scaling.</summary>
-        public bool useDynamicScale;
-        ///<summary>Memory less flag.</summary>
-        public RenderTextureMemoryless memoryless;
-        ///<summary>Texture name.</summary>
-        public string name;
-        ///<summary>Descriptor to determine how the texture will be in fast memory on platform that supports it.</summary>
-        public FastMemoryDesc fastMemoryDesc;
-
-        // Initial state. Those should not be used in the hash
-        ///<summary>Texture needs to be cleared on first use.</summary>
-        public bool clearBuffer;
-        ///<summary>Clear color.</summary>
-        public Color clearColor;
-
-        void InitDefaultValues(bool dynamicResolution, bool xrReady)
-        {
-            useDynamicScale = dynamicResolution;
-            // XR Ready
-            if (xrReady)
-            {
-                slices = TextureXR.slices;
-                dimension = TextureXR.dimension;
-            }
-            else
-            {
-                slices = 1;
-                dimension = TextureDimension.Tex2D;
-            }
-        }
-
-        /// <summary>
-        /// TextureDesc constructor for a texture using explicit size
-        /// </summary>
-        /// <param name="width">Texture width</param>
-        /// <param name="height">Texture height</param>
-        /// <param name="dynamicResolution">Use dynamic resolution</param>
-        /// <param name="xrReady">Set this to true if the Texture is a render texture in an XR setting.</param>
-        public TextureDesc(int width, int height, bool dynamicResolution = false, bool xrReady = false)
-            : this()
-        {
-            // Size related init
-            sizeMode = TextureSizeMode.Explicit;
-            this.width = width;
-            this.height = height;
-            // Important default values not handled by zero construction in this()
-            msaaSamples = MSAASamples.None;
-            InitDefaultValues(dynamicResolution, xrReady);
-        }
-
-        /// <summary>
-        /// TextureDesc constructor for a texture using a fixed scaling
-        /// </summary>
-        /// <param name="scale">RTHandle scale used for this texture</param>
-        /// <param name="dynamicResolution">Use dynamic resolution</param>
-        /// <param name="xrReady">Set this to true if the Texture is a render texture in an XR setting.</param>
-        public TextureDesc(Vector2 scale, bool dynamicResolution = false, bool xrReady = false)
-            : this()
-        {
-            // Size related init
-            sizeMode = TextureSizeMode.Scale;
-            this.scale = scale;
-            // Important default values not handled by zero construction in this()
-            msaaSamples = MSAASamples.None;
-            dimension = TextureDimension.Tex2D;
-            InitDefaultValues(dynamicResolution, xrReady);
-        }
-
-        /// <summary>
-        /// TextureDesc constructor for a texture using a functor for scaling
-        /// </summary>
-        /// <param name="func">Function used to determnine the texture size</param>
-        /// <param name="dynamicResolution">Use dynamic resolution</param>
-        /// <param name="xrReady">Set this to true if the Texture is a render texture in an XR setting.</param>
-        public TextureDesc(ScaleFunc func, bool dynamicResolution = false, bool xrReady = false)
-            : this()
-        {
-            // Size related init
-            sizeMode = TextureSizeMode.Functor;
-            this.func = func;
-            // Important default values not handled by zero construction in this()
-            msaaSamples = MSAASamples.None;
-            dimension = TextureDimension.Tex2D;
-            InitDefaultValues(dynamicResolution, xrReady);
-        }
-
-        /// <summary>
-        /// Copy constructor
-        /// </summary>
-        /// <param name="input"></param>
-        public TextureDesc(TextureDesc input)
-        {
-            this = input;
-        }
-
-        /// <summary>
-        /// Hash function
-        /// </summary>
-        /// <returns>The texture descriptor hash.</returns>
-        public override int GetHashCode()
-        {
-            int hashCode = 17;
-
-            unchecked
-            {
-                switch (sizeMode)
-                {
-                    case TextureSizeMode.Explicit:
-                        hashCode = hashCode * 23 + width;
-                        hashCode = hashCode * 23 + height;
-                        hashCode = hashCode * 23 + (int)msaaSamples;
-                        break;
-                    case TextureSizeMode.Functor:
-                        if (func != null)
-                            hashCode = hashCode * 23 + func.GetHashCode();
-                        hashCode = hashCode * 23 + (enableMSAA ? 1 : 0);
-                        break;
-                    case TextureSizeMode.Scale:
-                        hashCode = hashCode * 23 + scale.x.GetHashCode();
-                        hashCode = hashCode * 23 + scale.y.GetHashCode();
-                        hashCode = hashCode * 23 + (enableMSAA ? 1 : 0);
-                        break;
-                }
-
-                hashCode = hashCode * 23 + mipMapBias.GetHashCode();
-                hashCode = hashCode * 23 + slices;
-                hashCode = hashCode * 23 + (int)depthBufferBits;
-                hashCode = hashCode * 23 + (int)colorFormat;
-                hashCode = hashCode * 23 + (int)filterMode;
-                hashCode = hashCode * 23 + (int)wrapMode;
-                hashCode = hashCode * 23 + (int)dimension;
-                hashCode = hashCode * 23 + (int)memoryless;
-                hashCode = hashCode * 23 + anisoLevel;
-                hashCode = hashCode * 23 + (enableRandomWrite ? 1 : 0);
-                hashCode = hashCode * 23 + (useMipMap ? 1 : 0);
-                hashCode = hashCode * 23 + (autoGenerateMips ? 1 : 0);
-                hashCode = hashCode * 23 + (isShadowMap ? 1 : 0);
-                hashCode = hashCode * 23 + (bindTextureMS ? 1 : 0);
-                hashCode = hashCode * 23 + (useDynamicScale ? 1 : 0);
-            }
-
-            return hashCode;
-        }
-    }
-    #endregion
-
     /// <summary>
     /// The RenderGraphResourceRegistry holds all resource allocated during Render Graph execution.
     /// </summary>
@@ -329,41 +12,66 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
     {
         static readonly ShaderTagId s_EmptyName = new ShaderTagId("");
 
-        #region Resources
-        internal struct TextureResource
+        class IRenderGraphResource
         {
-            public TextureDesc  desc;
-            public bool         imported;
-            public RTHandle     rt;
-            public int          cachedHash;
-            public int          shaderProperty;
-            public bool         wasReleased;
+            public bool imported;
+            public int  cachedHash;
+            public int  shaderProperty;
+            public int  transientPassIndex;
+            public bool wasReleased;
 
-            internal TextureResource(RTHandle rt, int shaderProperty)
-                : this()
-            {
-                Reset();
-
-                this.rt = rt;
-                imported = true;
-                this.shaderProperty = shaderProperty;
-            }
-
-            internal TextureResource(in TextureDesc desc, int shaderProperty)
-                : this()
-            {
-                Reset();
-
-                this.desc = desc;
-                this.shaderProperty = shaderProperty;
-            }
-
-            void Reset()
+            public virtual void Reset()
             {
                 imported = false;
-                rt = null;
                 cachedHash = -1;
+                shaderProperty = 0;
+                transientPassIndex = -1;
                 wasReleased = false;
+            }
+
+            public virtual string GetName()
+            {
+                return "";
+            }
+        }
+
+        #region Resources
+        [DebuggerDisplay("Resource ({ResType}:{GetName()})")]
+        class RenderGraphResource<DescType, ResType>
+            : IRenderGraphResource
+            where DescType : struct
+            where ResType : class
+        {
+            public DescType desc;
+            public ResType  resource;
+
+            protected RenderGraphResource()
+            {
+
+            }
+
+            public override void Reset()
+            {
+                base.Reset();
+                resource = null;
+            }
+        }
+
+        [DebuggerDisplay("TextureResource ({desc.name})")]
+        class TextureResource : RenderGraphResource<TextureDesc, RTHandle>
+        {
+            public override string GetName()
+            {
+                return desc.name;
+            }
+        }
+
+        [DebuggerDisplay("ComputeBufferResource ({desc.name})")]
+        class ComputeBufferResource : RenderGraphResource<ComputeBufferDesc, ComputeBuffer>
+        {
+            public override string GetName()
+            {
+                return desc.name;
             }
         }
 
@@ -379,31 +87,19 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
             }
         }
 
-        internal struct ComputeBufferResource
-        {
-            public ComputeBuffer    computeBuffer;
-            public bool             imported;
-
-            internal ComputeBufferResource(ComputeBuffer computeBuffer, bool imported)
-            {
-                this.computeBuffer = computeBuffer;
-                this.imported = imported;
-            }
-        }
         #endregion
 
-        DynamicArray<TextureResource>       m_TextureResources = new DynamicArray<TextureResource>();
-        Dictionary<int, Stack<RTHandle>>    m_TexturePool = new Dictionary<int, Stack<RTHandle>>();
+        DynamicArray<IRenderGraphResource>[] m_Resources = new DynamicArray<IRenderGraphResource>[(int)RenderGraphResourceType.Count];
+
+        TexturePool                         m_TexturePool = new TexturePool();
+        ComputeBufferPool                   m_ComputeBufferPool = new ComputeBufferPool();
         DynamicArray<RendererListResource>  m_RendererListResources = new DynamicArray<RendererListResource>();
-        DynamicArray<ComputeBufferResource> m_ComputeBufferResources = new DynamicArray<ComputeBufferResource>();
         RTHandleSystem                      m_RTHandleSystem = new RTHandleSystem();
         RenderGraphDebugParams              m_RenderGraphDebug;
         RenderGraphLogger                   m_Logger;
+        int                                 m_CurrentFrameIndex;
 
         RTHandle                            m_CurrentBackbuffer;
-
-        // Diagnostic only
-        List<(int, RTHandle)>               m_AllocatedTextures = new List<(int, RTHandle)>();
 
         #region Public Interface
         /// <summary>
@@ -416,17 +112,7 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
             if (!handle.IsValid())
                 return null;
 
-#if DEVELOPMENT_BUILD || UNITY_EDITOR
-
-            var res = m_TextureResources[handle];
-
-            if (res.rt == null && !res.wasReleased)
-                throw new InvalidOperationException(string.Format("Trying to access texture \"{0}\" that was never created. Check that it was written at least once before trying to get it.", res.desc.name));
-
-            if (res.rt == null && res.wasReleased)
-                throw new InvalidOperationException(string.Format("Trying to access texture \"{0}\" that was already released. Check that the last pass where it's read is after this one.", res.desc.name));
-#endif
-            return m_TextureResources[handle].rt;
+            return GetTextureResource(handle.handle).resource;
         }
 
         /// <summary>
@@ -452,7 +138,7 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
             if (!handle.IsValid())
                 return null;
 
-            return m_ComputeBufferResources[handle].computeBuffer;
+            return GetComputeBufferResource(handle.handle).resource;
         }
         #endregion
 
@@ -468,25 +154,70 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
             m_RTHandleSystem.Initialize(Screen.width, Screen.height, supportMSAA, initialSampleCount);
             m_RenderGraphDebug = renderGraphDebug;
             m_Logger = logger;
+
+            for (int i = 0; i < (int)RenderGraphResourceType.Count; ++i)
+                m_Resources[i] = new DynamicArray<IRenderGraphResource>();
         }
 
-        internal void SetRTHandleReferenceSize(int width, int height, MSAASamples msaaSamples)
+        ResType GetResource<DescType, ResType>(DynamicArray<IRenderGraphResource> resourceArray, int index)
+            where DescType : struct
+            where ResType : class
         {
+            var res = resourceArray[index] as RenderGraphResource<DescType, ResType>;
+
+#if DEVELOPMENT_BUILD || UNITY_EDITOR
+            if (res.resource == null && !res.wasReleased)
+                throw new InvalidOperationException(string.Format("Trying to access resource \"{0}\" that was never created. Check that it was written at least once before trying to get it.", res.GetName()));
+
+            if (res.resource == null && res.wasReleased)
+                throw new InvalidOperationException(string.Format("Trying to access resource \"{0}\" that was already released. Check that the last pass where it's read is after this one.", res.GetName()));
+#endif
+            return res.resource;
+        }
+
+        internal void BeginRender(int width, int height, MSAASamples msaaSamples, int currentFrameIndex)
+        {
+            m_CurrentFrameIndex = currentFrameIndex;
+            // Update RTHandleSystem with size for this rendering pass.
             m_RTHandleSystem.SetReferenceSize(width, height, msaaSamples);
         }
 
         internal RTHandleProperties GetRTHandleProperties() { return m_RTHandleSystem.rtHandleProperties; }
 
+        void CheckHandleValidity(in ResourceHandle res)
+        {
+            var resources = m_Resources[res.iType];
+            if (res.index >= resources.size)
+                throw new ArgumentException($"Trying to access resource of type {res.type} with an invalid resource index {res.index}");
+        }
+
+        internal string GetResourceName(in ResourceHandle res)
+        {
+            CheckHandleValidity(res);
+            return m_Resources[res.iType][res.index].GetName();
+        }
+
+        internal bool IsResourceImported(in ResourceHandle res)
+        {
+            CheckHandleValidity(res);
+            return m_Resources[res.iType][res.index].imported;
+        }
+
+        internal int GetResourceTransientIndex(in ResourceHandle res)
+        {
+            CheckHandleValidity(res);
+            return m_Resources[res.iType][res.index].transientPassIndex;
+        }
+
         // Texture Creation/Import APIs are internal because creation should only go through RenderGraph
         internal TextureHandle ImportTexture(RTHandle rt, int shaderProperty = 0)
         {
-            int newHandle = m_TextureResources.Add(new TextureResource(rt, shaderProperty));
-            return new TextureHandle(newHandle);
-        }
+            int newHandle = AddNewResource(m_Resources[(int)RenderGraphResourceType.Texture], out TextureResource texResource);
+            texResource.resource = rt;
+            texResource.imported = true;
+            texResource.shaderProperty = shaderProperty;
 
-        internal bool IsTextureImported(TextureHandle handle)
-        {
-            return handle.IsValid() ? GetTextureResource(handle).imported : false;
+            return new TextureHandle(newHandle);
         }
 
         internal TextureHandle ImportBackbuffer(RenderTargetIdentifier rt)
@@ -496,31 +227,50 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
             else
                 m_CurrentBackbuffer = m_RTHandleSystem.Alloc(rt);
 
-            int newHandle = m_TextureResources.Add(new TextureResource(m_CurrentBackbuffer, 0));
+            int newHandle = AddNewResource(m_Resources[(int)RenderGraphResourceType.Texture], out TextureResource texResource);
+            texResource.resource = m_CurrentBackbuffer;
+            texResource.imported = true;
+
             return new TextureHandle(newHandle);
+        }
+
+        int AddNewResource<ResType>(DynamicArray<IRenderGraphResource> resourceArray, out ResType outRes) where ResType : IRenderGraphResource, new()
+        {
+            // In order to not create garbage, instead of using Add, we keep the content of the array while resizing and we just reset the existing ref (or create it if it's null).
+            int result = resourceArray.size;
+            resourceArray.Resize(resourceArray.size + 1, true);
+            if (resourceArray[result] == null)
+                resourceArray[result] = new ResType();
+
+            outRes = resourceArray[result] as ResType;
+            outRes.Reset();
+            return result;
         }
 
         internal TextureHandle CreateTexture(in TextureDesc desc, int shaderProperty = 0, int transientPassIndex = -1)
         {
             ValidateTextureDesc(desc);
 
-            int newHandle = m_TextureResources.Add(new TextureResource(desc, shaderProperty));
-            return new TextureHandle(newHandle, transientPassIndex);
+            int newHandle = AddNewResource(m_Resources[(int)RenderGraphResourceType.Texture], out TextureResource texResource);
+            texResource.desc = desc;
+            texResource.shaderProperty = shaderProperty;
+            texResource.transientPassIndex = transientPassIndex;
+            return new TextureHandle(newHandle);
         }
 
         internal int GetTextureResourceCount()
         {
-            return m_TextureResources.size;
+            return m_Resources[(int)RenderGraphResourceType.Texture].size;
         }
 
-        ref TextureResource GetTextureResource(TextureHandle res)
+        TextureResource GetTextureResource(in ResourceHandle handle)
         {
-            return ref m_TextureResources[res];
+            return m_Resources[(int)RenderGraphResourceType.Texture][handle] as TextureResource;
         }
 
-        internal TextureDesc GetTextureResourceDesc(TextureHandle res)
+        internal TextureDesc GetTextureResourceDesc(in ResourceHandle handle)
         {
-            return m_TextureResources[res].desc;
+            return (m_Resources[(int)RenderGraphResourceType.Texture][handle] as TextureResource).desc;
         }
 
         internal RendererListHandle CreateRendererList(in RendererListDesc desc)
@@ -533,37 +283,90 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
 
         internal ComputeBufferHandle ImportComputeBuffer(ComputeBuffer computeBuffer)
         {
-            int newHandle = m_ComputeBufferResources.Add(new ComputeBufferResource(computeBuffer, imported: true));
+            int newHandle = AddNewResource(m_Resources[(int)RenderGraphResourceType.ComputeBuffer], out ComputeBufferResource bufferResource);
+            bufferResource.resource = computeBuffer;
+            bufferResource.imported = true;
+
             return new ComputeBufferHandle(newHandle);
         }
 
-        internal bool IsComputeBufferImported(ComputeBufferHandle handle)
+        internal ComputeBufferHandle CreateComputeBuffer(in ComputeBufferDesc desc, int transientPassIndex = -1)
         {
-            return handle.IsValid() ? GetComputeBufferResource(handle).imported : false;
+            ValidateComputeBufferDesc(desc);
+
+            int newHandle = AddNewResource(m_Resources[(int)RenderGraphResourceType.ComputeBuffer], out ComputeBufferResource bufferResource);
+            bufferResource.desc = desc;
+            bufferResource.transientPassIndex = transientPassIndex;
+
+            return new ComputeBufferHandle(newHandle);
+        }
+
+        internal ComputeBufferDesc GetComputeBufferResourceDesc(in ResourceHandle handle)
+        {
+            return (m_Resources[(int)RenderGraphResourceType.ComputeBuffer][handle] as ComputeBufferResource).desc;
         }
 
         internal int GetComputeBufferResourceCount()
         {
-            return m_ComputeBufferResources.size;
+            return m_Resources[(int)RenderGraphResourceType.ComputeBuffer].size;
         }
 
-        internal ref ComputeBufferResource GetComputeBufferResource(ComputeBufferHandle res)
+        ComputeBufferResource GetComputeBufferResource(in ResourceHandle handle)
         {
-            return ref m_ComputeBufferResources[res];
+            return m_Resources[(int)RenderGraphResourceType.ComputeBuffer][handle] as ComputeBufferResource;
         }
 
-        internal void CreateAndClearTexture(RenderGraphContext rgContext, TextureHandle texture)
+        internal void CreateAndClearTexture(RenderGraphContext rgContext, int index)
         {
-            ref var resource = ref GetTextureResource(texture);
+            var resource = m_Resources[(int)RenderGraphResourceType.Texture][index] as TextureResource;
+
             if (!resource.imported)
             {
-                CreateTextureForPass(ref resource);
+                var desc = resource.desc;
+                int hashCode = desc.GetHashCode();
 
+                if (resource.resource != null)
+                    throw new InvalidOperationException(string.Format("Trying to create an already created texture ({0}). Texture was probably declared for writing more than once in the same pass.", resource.desc.name));
+
+                resource.resource = null;
+                if (!m_TexturePool.TryGetResource(hashCode, out resource.resource))
+                {
+                    string name = desc.name;
+                    if (m_RenderGraphDebug.tagResourceNamesWithRG)
+                        name = $"RenderGraph_{name}";
+
+                    // Note: Name used here will be the one visible in the memory profiler so it means that whatever is the first pass that actually allocate the texture will set the name.
+                    // TODO: Find a way to display name by pass.
+                    switch (desc.sizeMode)
+                    {
+                        case TextureSizeMode.Explicit:
+                            resource.resource = m_RTHandleSystem.Alloc(desc.width, desc.height, desc.slices, desc.depthBufferBits, desc.colorFormat, desc.filterMode, desc.wrapMode, desc.dimension, desc.enableRandomWrite,
+                            desc.useMipMap, desc.autoGenerateMips, desc.isShadowMap, desc.anisoLevel, desc.mipMapBias, desc.msaaSamples, desc.bindTextureMS, desc.useDynamicScale, desc.memoryless, desc.name);
+                            break;
+                        case TextureSizeMode.Scale:
+                            resource.resource = m_RTHandleSystem.Alloc(desc.scale, desc.slices, desc.depthBufferBits, desc.colorFormat, desc.filterMode, desc.wrapMode, desc.dimension, desc.enableRandomWrite,
+                            desc.useMipMap, desc.autoGenerateMips, desc.isShadowMap, desc.anisoLevel, desc.mipMapBias, desc.enableMSAA, desc.bindTextureMS, desc.useDynamicScale, desc.memoryless, desc.name);
+                            break;
+                        case TextureSizeMode.Functor:
+                            resource.resource = m_RTHandleSystem.Alloc(desc.func, desc.slices, desc.depthBufferBits, desc.colorFormat, desc.filterMode, desc.wrapMode, desc.dimension, desc.enableRandomWrite,
+                            desc.useMipMap, desc.autoGenerateMips, desc.isShadowMap, desc.anisoLevel, desc.mipMapBias, desc.enableMSAA, desc.bindTextureMS, desc.useDynamicScale, desc.memoryless, desc.name);
+                            break;
+                    }
+                }
+
+                //// Try to update name when re-using a texture.
+                //// TODO RENDERGRAPH: Check if that actually works.
+                //resource.rt.name = desc.name;
+
+                resource.cachedHash = hashCode;
+
+#if UNITY_2020_2_OR_NEWER
                 var fastMemDesc = resource.desc.fastMemoryDesc;
                 if(fastMemDesc.inFastMemory)
                 {
-                    resource.rt.SwitchToFastMemory(rgContext.cmd, fastMemDesc.residencyFraction, fastMemDesc.flags);
+                    resource.resource.SwitchToFastMemory(rgContext.cmd, fastMemDesc.residencyFraction, fastMemDesc.flags);
                 }
+#endif
 
                 if (resource.desc.clearBuffer || m_RenderGraphDebug.clearRenderTargetsAtCreation)
                 {
@@ -573,119 +376,110 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
                     {
                         var clearFlag = resource.desc.depthBufferBits != DepthBits.None ? ClearFlag.Depth : ClearFlag.Color;
                         var clearColor = debugClear ? Color.magenta : resource.desc.clearColor;
-                        CoreUtils.SetRenderTarget(rgContext.cmd, resource.rt, clearFlag, clearColor);
+                        CoreUtils.SetRenderTarget(rgContext.cmd, resource.resource, clearFlag, clearColor);
                     }
                 }
 
-                LogTextureCreation(resource.rt, resource.desc.clearBuffer || m_RenderGraphDebug.clearRenderTargetsAtCreation);
+                m_TexturePool.RegisterFrameAllocation(hashCode, resource.resource);
+                LogTextureCreation(resource.resource, resource.desc.clearBuffer || m_RenderGraphDebug.clearRenderTargetsAtCreation);
             }
         }
 
-        void CreateTextureForPass(ref TextureResource resource)
+        internal void CreateComputeBuffer(RenderGraphContext rgContext, int index)
         {
-            var desc = resource.desc;
-            int hashCode = desc.GetHashCode();
-
-            if(resource.rt != null)
-                throw new InvalidOperationException(string.Format("Trying to create an already created texture ({0}). Texture was probably declared for writing more than once in the same pass.", resource.desc.name));
-
-            resource.rt = null;
-            if (!TryGetRenderTarget(hashCode, out resource.rt))
+            var resource = m_Resources[(int)RenderGraphResourceType.ComputeBuffer][index] as ComputeBufferResource;
+            if (!resource.imported)
             {
-                // Note: Name used here will be the one visible in the memory profiler so it means that whatever is the first pass that actually allocate the texture will set the name.
-                // TODO: Find a way to display name by pass.
-                switch (desc.sizeMode)
+                var desc = resource.desc;
+                int hashCode = desc.GetHashCode();
+
+                if (resource.resource != null)
+                    throw new InvalidOperationException(string.Format("Trying to create an already created Compute Buffer ({0}). Buffer was probably declared for writing more than once in the same pass.", resource.desc.name));
+
+                resource.resource = null;
+                if (!m_ComputeBufferPool.TryGetResource(hashCode, out resource.resource))
                 {
-                    case TextureSizeMode.Explicit:
-                        resource.rt = m_RTHandleSystem.Alloc(desc.width, desc.height, desc.slices, desc.depthBufferBits, desc.colorFormat, desc.filterMode, desc.wrapMode, desc.dimension, desc.enableRandomWrite,
-                        desc.useMipMap, desc.autoGenerateMips, desc.isShadowMap, desc.anisoLevel, desc.mipMapBias, desc.msaaSamples, desc.bindTextureMS, desc.useDynamicScale, desc.memoryless, desc.name);
-                        break;
-                    case TextureSizeMode.Scale:
-                        resource.rt = m_RTHandleSystem.Alloc(desc.scale, desc.slices, desc.depthBufferBits, desc.colorFormat, desc.filterMode, desc.wrapMode, desc.dimension, desc.enableRandomWrite,
-                        desc.useMipMap, desc.autoGenerateMips, desc.isShadowMap, desc.anisoLevel, desc.mipMapBias, desc.enableMSAA, desc.bindTextureMS, desc.useDynamicScale, desc.memoryless, desc.name);
-                        break;
-                    case TextureSizeMode.Functor:
-                        resource.rt = m_RTHandleSystem.Alloc(desc.func, desc.slices, desc.depthBufferBits, desc.colorFormat, desc.filterMode, desc.wrapMode, desc.dimension, desc.enableRandomWrite,
-                        desc.useMipMap, desc.autoGenerateMips, desc.isShadowMap, desc.anisoLevel, desc.mipMapBias, desc.enableMSAA, desc.bindTextureMS, desc.useDynamicScale, desc.memoryless, desc.name);
-                        break;
+                    resource.resource = new ComputeBuffer(resource.desc.count, resource.desc.stride, resource.desc.type);
+                    resource.resource.name = m_RenderGraphDebug.tagResourceNamesWithRG ? $"RenderGraph_{resource.desc.name}" : resource.desc.name;
                 }
+                resource.cachedHash = hashCode;
+
+                m_ComputeBufferPool.RegisterFrameAllocation(hashCode, resource.resource);
+                LogComputeBufferCreation(resource.resource);
             }
-
-            //// Try to update name when re-using a texture.
-            //// TODO: Check if that actually works.
-            //resource.rt.name = desc.name;
-
-#if DEVELOPMENT_BUILD || UNITY_EDITOR
-            if (hashCode != -1)
-            {
-                m_AllocatedTextures.Add((hashCode, resource.rt));
-            }
-#endif
-
-            resource.cachedHash = hashCode;
         }
 
-        void SetGlobalTextures(RenderGraphContext rgContext, List<TextureHandle> textures, bool bindDummyTexture)
+        void SetGlobalTextures(RenderGraphContext rgContext, List<ResourceHandle> textures, bool bindDummyTexture)
         {
             foreach (var resource in textures)
             {
                 var resourceDesc = GetTextureResource(resource);
                 if (resourceDesc.shaderProperty != 0)
                 {
-                    if (resourceDesc.rt != null)
+                    if (resourceDesc.resource != null)
                     {
-                        rgContext.cmd.SetGlobalTexture(resourceDesc.shaderProperty, bindDummyTexture ? TextureXR.GetMagentaTexture() : resourceDesc.rt);
+                        rgContext.cmd.SetGlobalTexture(resourceDesc.shaderProperty, bindDummyTexture ? TextureXR.GetMagentaTexture() : resourceDesc.resource);
                     }
                 }
             }
         }
 
 
-        internal void PreRenderPassSetGlobalTextures(RenderGraphContext rgContext, List<TextureHandle> textures)
+        internal void PreRenderPassSetGlobalTextures(RenderGraphContext rgContext, List<ResourceHandle> textures)
         {
             SetGlobalTextures(rgContext, textures, false);
         }
 
-        internal void PostRenderPassUnbindGlobalTextures(RenderGraphContext rgContext, List<TextureHandle> textures)
+        internal void PostRenderPassUnbindGlobalTextures(RenderGraphContext rgContext, List<ResourceHandle> textures)
         {
             SetGlobalTextures(rgContext, textures, true);
         }
 
-        internal void ReleaseTexture(RenderGraphContext rgContext, TextureHandle resource)
+        internal void ReleaseTexture(RenderGraphContext rgContext, int index)
         {
-            ref var resourceDesc = ref GetTextureResource(resource);
-            if (!resourceDesc.imported)
+            var resource = m_Resources[(int)RenderGraphResourceType.Texture][index] as TextureResource;
+
+            if (!resource.imported)
             {
+                if (resource.resource == null)
+                    throw new InvalidOperationException($"Tried to release a texture ({resource.desc.name}) that was never created. Check that there is at least one pass writing to it first.");
+
                 if (m_RenderGraphDebug.clearRenderTargetsAtRelease)
                 {
                     using (new ProfilingScope(rgContext.cmd, ProfilingSampler.Get(RenderGraphProfileId.RenderGraphClearDebug)))
                     {
-                        var clearFlag = resourceDesc.desc.depthBufferBits != DepthBits.None ? ClearFlag.Depth : ClearFlag.Color;
-                        CoreUtils.SetRenderTarget(rgContext.cmd, GetTexture(resource), clearFlag, Color.magenta);
+                        var clearFlag = resource.desc.depthBufferBits != DepthBits.None ? ClearFlag.Depth : ClearFlag.Color;
+                        // Not ideal to do new TextureHandle here but GetTexture is a public API and we rather have it take an explicit TextureHandle parameters.
+                        // Everywhere else internally int is better because it allows us to share more code.
+                        CoreUtils.SetRenderTarget(rgContext.cmd, GetTexture(new TextureHandle(index)), clearFlag, Color.magenta);
                     }
                 }
 
-                LogTextureRelease(resourceDesc.rt);
-                ReleaseTextureResource(resourceDesc.cachedHash, resourceDesc.rt);
-                resourceDesc.cachedHash = -1;
-                resourceDesc.rt = null;
-                resourceDesc.wasReleased = true;
+                LogTextureRelease(resource.resource);
+                m_TexturePool.ReleaseResource(resource.cachedHash, resource.resource, m_CurrentFrameIndex);
+                m_TexturePool.UnregisterFrameAllocation(resource.cachedHash, resource.resource);
+                resource.cachedHash = -1;
+                resource.resource = null;
+                resource.wasReleased = true;
             }
         }
 
-        void ReleaseTextureResource(int hash, RTHandle rt)
+        internal void ReleaseComputeBuffer(RenderGraphContext rgContext, int index)
         {
-            if (!m_TexturePool.TryGetValue(hash, out var stack))
+            var resource = m_Resources[(int)RenderGraphResourceType.ComputeBuffer][index] as ComputeBufferResource;
+
+            if (!resource.imported)
             {
-                stack = new Stack<RTHandle>();
-                m_TexturePool.Add(hash, stack);
+                if (resource.resource == null)
+                    throw new InvalidOperationException($"Tried to release a compute buffer ({resource.desc.name}) that was never created. Check that there is at least one pass writing to it first.");
+
+                LogComputeBufferRelease(resource.resource);
+                m_ComputeBufferPool.ReleaseResource(resource.cachedHash, resource.resource, m_CurrentFrameIndex);
+                m_ComputeBufferPool.UnregisterFrameAllocation(resource.cachedHash, resource.resource);
+                resource.cachedHash = -1;
+                resource.resource = null;
+                resource.wasReleased = true;
             }
-
-            stack.Push(rt);
-
-#if DEVELOPMENT_BUILD || UNITY_EDITOR
-            m_AllocatedTextures.Remove((hash, rt));
-#endif
         }
 
         void ValidateTextureDesc(in TextureDesc desc)
@@ -744,16 +538,19 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
 #endif
         }
 
-        bool TryGetRenderTarget(int hashCode, out RTHandle rt)
+        void ValidateComputeBufferDesc(in ComputeBufferDesc desc)
         {
-            if (m_TexturePool.TryGetValue(hashCode, out var stack) && stack.Count > 0)
+#if DEVELOPMENT_BUILD || UNITY_EDITOR
+            // TODO RENDERGRAPH: Check actual condition on stride.
+            if (desc.stride % 4 != 0)
             {
-                rt = stack.Pop();
-                return true;
+                throw new ArgumentException("Invalid Compute Buffer creation descriptor: Compute Buffer stride must be at least 4.");
             }
-
-            rt = null;
-            return false;
+            if (desc.count == 0)
+            {
+                throw new ArgumentException("Invalid Compute Buffer creation descriptor: Compute Buffer count  must be non zero.");
+            }
+#endif
         }
 
         internal void CreateRendererLists(List<RendererListHandle> rendererLists)
@@ -773,29 +570,18 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
         {
             LogResources();
 
-            m_TextureResources.Clear();
+            for (int i = 0; i < (int)RenderGraphResourceType.Count; ++i)
+                m_Resources[i].Clear();
             m_RendererListResources.Clear();
-            m_ComputeBufferResources.Clear();
 
-#if DEVELOPMENT_BUILD || UNITY_EDITOR
-            if (m_AllocatedTextures.Count != 0 && !onException)
-            {
-                string logMessage = "RenderGraph: Not all textures were released.";
+            m_TexturePool.CheckFrameAllocation(onException, m_CurrentFrameIndex);
+            m_ComputeBufferPool.CheckFrameAllocation(onException, m_CurrentFrameIndex);
 
-                List<(int, RTHandle)> tempList = new List<(int, RTHandle)>(m_AllocatedTextures);
-                foreach (var value in tempList)
-                {
-                    logMessage = $"{logMessage}\n\t{value.Item2.name}";
-                    ReleaseTextureResource(value.Item1, value.Item2);
-                }
-
-                Debug.LogWarning(logMessage);
-            }
-
-            // If an error occurred during execution, it's expected that textures are not all release so we clear the trakcing list.
-            if (onException)
-                m_AllocatedTextures.Clear();
-#endif
+            // TODO RENDERGRAPH: Might not be ideal to purge stale resources every frame.
+            // In case users enable/disable features along a level it might provoke performance spikes when things are reallocated...
+            // Will be much better when we have actual resource aliasing and we can manage memory more efficiently.
+            m_TexturePool.PurgeUnusedResources(m_CurrentFrameIndex);
+            m_ComputeBufferPool.PurgeUnusedResources(m_CurrentFrameIndex);
         }
 
         internal void ResetRTHandleReferenceSize(int width, int height)
@@ -805,20 +591,15 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
 
         internal void Cleanup()
         {
-            foreach (var value in m_TexturePool)
-            {
-                foreach (var rt in value.Value)
-                {
-                    m_RTHandleSystem.Release(rt);
-                }
-            }
+            m_TexturePool.Cleanup();
+            m_ComputeBufferPool.Cleanup();
         }
 
         void LogTextureCreation(RTHandle rt, bool cleared)
         {
             if (m_RenderGraphDebug.logFrameInformation)
             {
-                m_Logger.LogLine("Created Texture: {0} (Cleared: {1})", rt.rt.name, cleared);
+                m_Logger.LogLine($"Created Texture: {rt.rt.name} (Cleared: {cleared})");
             }
         }
 
@@ -826,7 +607,23 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
         {
             if (m_RenderGraphDebug.logFrameInformation)
             {
-                m_Logger.LogLine("Released Texture: {0}", rt.rt.name);
+                m_Logger.LogLine($"Released Texture: {rt.rt.name}");
+            }
+        }
+
+        void LogComputeBufferCreation(ComputeBuffer buffer)
+        {
+            if (m_RenderGraphDebug.logFrameInformation)
+            {
+                m_Logger.LogLine($"Created ComputeBuffer: {buffer}");
+            }
+        }
+
+        void LogComputeBufferRelease(ComputeBuffer buffer)
+        {
+            if (m_RenderGraphDebug.logFrameInformation)
+            {
+                m_Logger.LogLine($"Released ComputeBuffer: {buffer}");
             }
         }
 
@@ -836,19 +633,8 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
             {
                 m_Logger.LogLine("==== Allocated Resources ====\n");
 
-                List<string> allocationList = new List<string>();
-                foreach (var stack in m_TexturePool)
-                {
-                    foreach (var rt in stack.Value)
-                    {
-                        allocationList.Add(rt.rt.name);
-                    }
-                }
-
-                allocationList.Sort();
-                int index = 0;
-                foreach (var element in allocationList)
-                    m_Logger.LogLine("[{0}] {1}", index++, element);
+                m_TexturePool.LogResources(m_Logger);
+                m_ComputeBufferPool.LogResources(m_Logger);
             }
         }